Add logging of Rails request ID
The request ID was introduced in Rails 3.2 to enable easy tracking of
individual requests in logs. This adds it to the Request tab in bugsnag,
to make tracking down logs for a particular error easier.

The ID is generated by ActionDispatch::RequestId (using either the
`X-Request-Id` header, or `SecureRandom.uuid`).