Re-init delivery thread if current PID does not match the one it was created with
Currently, a delivery thread is started one time per process using the
start_once! method.

However, this causes problems in forking code.  If a process encounters
an exception and notifies, then forks, deliveries in the child process
will not be delivered since it will not be given a delivery thread since @started
will be true (inherited from parent process).

This commit modifies the start_once! method to store the PID of the
process in @started.  Then, instead of just checking if @started is set,
it checks if it matched the current PID.  If it doesn't match, the
current process doesn't have a delivery thread so we create one.