Adding a flag on Bugsnag::Railtie to control app_type overwriting
Sometimes rails is ran as a dependency of another integration (eg. Sidekiq).
But since the rails initializer runs after the integration, the app_type that
is normally set ends up overwritten.

Each integration has it's own logic that determines if rails is a dependency
or the main actor. This flag can now be set by them to make sure the
app_type won't change.