Trim nested arrays and drop metadata to save on payload size (#295)
If a payload contains an array of max length strings (3kb), trimming to
40 lines leaves 5kb for the rest of the payload including a long message
and API key, et cetera. The only concern is losing debugging information
to a smarter heuristic than truncating the lines outside the allowed
length.

In the (extreme) case that trimming strings and array truncation does
not result in a payload of appropriate size, remove the `metaData` key
from an array of events. The error will at least be reported, even if 
custom metadata is unavailable to assist with debugging.