Set the rails logger after initialisation
Currently, bugsnag logging does not work in rails unless you explicitly set the logger in your bugsnag initializer, and so all logging messages are being lost. This seems to be because `Rails.logger` has not been set up yet at the point when bugsnag is accessing it. Setting the logger after initialization fixes the issue.