Improve code extraction performance
We now open each file in a stacktrace once to save opening the same file
multiple times if it is in the stacktrace multiple times. It's quite
common to have the same file in multiple stacktrace lines, e.g. when
a private method raises. The performance gain is especially noticeable
when a a really long file is repeated in the stacktrace as we have to
iterate through each line to find the ones we're interested in

e.g. a trace may look like this:

'/some/file line 5 in "foo"'
'/some/file line 10 in "bar"'
'/some/file line 15 in "baz"'

Previously we would:
1. Open the file for frame 1
2. Read each line up to line 11
3. Trim lines down to lines 2-8
4. Open the file for frame 2
5. Read each line up to line 14
6. Trim lines down to lines 7-13
7. Open the file for frame 3
8. Read each line up to line 19
9. Trim lines down to lines 12-18

Now we:
1. Open the file once
2. Read from line 1 until line 19
3. Add the lines each frame cares about:
    a. Lines 1-11 for frame 1
    b. Lines 3-14 for frame 2
    c. Lines 8-19 for frame 3
3. Trim lines down:
    a. To lines 2-8 for frame 1
    b. To lines 7-13 for frame 2
    c. To lines 12-18 for frame 3

Trimming is necessary because we need to record more lines than we
actually want to keep, so that we can always record 7 lines of context,
even when the frame starts at line 1 or the last line in the file

The new way is still quite inefficient and requires iterating over the
stack frames multiple times, however this passes all tests and can be
improved after profiling it against the old method