Make Cleaner a shared instance
This exposed a pretty big regression where filters wouldn't match when
they should have. This was caused by us filtering the entire report
object in one go, which means the scopes were nested deeper than they
were before

Previously we filtered the events.metaData directly, so scopes would not
include 'events.metaData' and therefore a filter of 'foo' would match
'events.metaData.foo'. Now that we filter the entire report, if a filter
relied on 'deep_filters', it would apply and so things that should be
redacted wouldn't have been

To solve this, we strip each 'scope_to_filter' from the scope before
matching it, if deep_filters are enabled

The tests passed before this change because we set 'scopes_to_filter'
in each test. Now that the instance is shared, the scopes are fetched
from the Configuration so this isn't possible and it exposed the bug

The tests now cover this case, because they can't set 'scopes_to_filter'
directly anymore, so they are testing that the filters they're using
match with the Configuration's 'scopes_to_filter'