Extract a helper to fetch the code from an event
This also uses 'to_a' to ensure the order of code is correct. Hashes are
compared without regard for order because usually the order doesn't
matter. I don't think the API cares about the order because we explicitly
number the lines, but it's useful to validate our code is doing what is
expected