Add 'on_error' callbacks, backed by middleware
These new 'on_error' callbacks will ultimately be a replacement for
the existing 'before_notify_callbacks'

The main problem these solve is that they are global, rather than
being scoped to the current thread. This means you can add a callback
in a Rails initializer and it will run as expected. With the current
'before_notify_callbacks', they only run in the same thread as they
are added and so won't run if created in an initializer. This is a
problem in most (all ?) frameworks, not just Rails

This adds some complexity to the MiddlewareStack, because it now has
to handle procs being passed as middleware. It's important that we
don't wrap the procs earlier for two reasons:

  1. We need to be able to remove them, which is much simpler if
     we can do this by object identity (because the current method
       works that way)
  2. We need to be able to abort the 'queue' of callbacks, if one
     returns false. This is much easier if we have a list of procs
     to run