Add test with exit after an exception
This proves that we don't unwrap exceptions too much when unwrapping
Bundler's friendly errors. In this case, the backtrace will look
something like this:

1. Bundler friendly errors
2. Our exit
3. The exception we raised

We want to drop 1. because it may be hiding a real exception, however
we don't want to drop 2. because it's a deliberate exit

This test proves that we don't go too far and end up with an exception
that we would notify about