Add Note for installing test group gems.
Added a note that in order to run the specs, you may need to install the bundle with the `--with test` command line option in order to install the `test` group of gems, including `rake`. Otherwise you end up with an error like this: 

```bash
$ be rake spec


bundler: failed to load command: rake (/Users/joshuapinter/.rbenv/versions/2.6.5/bin/rake)
Gem::Exception: can't find executable rake for gem rake. rake is not currently included in the bundle, perhaps you meant to add it to your Gemfile?
```

Wasted 15 minutes figuring out what the issue was so this should help others contribute.