Allow thread_queue to restart if needed
In cases where Bugsnag's at_exit runs _after_ the thread_queue's
at_exit, the payload will never be delivered. This is because we
stopped the thread queue from running and didn't allow it to restart

With this change we allow the thread queue to be restarted after it
has finished (the worker thread is done). This will happen automatically
when a new payload needs to be sent