Fix possible NoMethodError in railtie
This can happen if Bugsnag is added to the Gemfile with 'require: false'.
In this case, 'require "bugsnag"' in a Rails initializer will run the
'config.before_initialize' before the rest of the Railtie has been
executed. This results in the 'event_subscription' method being
undefined as it was declared below the line that calls it

By moving it to the top of the file this is no longer a problem